Added testing for our request method overrides
